+++ lib/sentry/sources.ex
@@ -9,8 +9,64 @@ defmodule Sentry.Sources do
           }
         }
 
+  @source_code_map_key {:sentry, :source_code_map}
+
+  @spec load_source_code_map_if_present() :: :ok
+  def load_source_code_map_if_present do
+    path = Path.relative_to_cwd(path_of_packaged_source_code())
+
+    with {:ok, contents} <- File.read(path),
+         {:ok, source_map} <- decode_source_code_map(contents) do
+      :persistent_term.put(@source_code_map_key, source_map)
+    else
+      {:error, :binary_to_term} ->
+        IO.warn("""
+        Sentry found a source code map file at #{path}, but it was unable to decode its
+        contents.
+        """)
+
+      {:error, :enoent} ->
+        :ok
+
+      {:error, reason} ->
+        IO.warn("""
+        Sentry found a source code map file at #{path}, but it was unable to read it.
+        The reason was: #{:file.format_error(reason)}
+        """)
+    end
+
+    :ok
+  end
+
+  @spec path_of_packaged_source_code() :: Path.t()
+  def path_of_packaged_source_code do
+    Path.join([Application.app_dir(:sentry), "priv", "sentry.map"])
+  end
+
+  @spec encode_source_code_map(source_map()) :: binary()
+  def encode_source_code_map(%{} = source_map) do
+    # This term contains no atoms, so that it can be decoded with binary_to_term(bin, [:safe]).
+    term_to_encode = %{"version" => 1, "files_map" => source_map}
+    :erlang.term_to_binary(term_to_encode, compressed: 9)
+  end
+
+  defp decode_source_code_map(binary) when is_binary(binary) do
+    try do
+      :erlang.binary_to_term(binary, [:safe])
+    rescue
+      ArgumentError -> {:error, :binary_to_term}
+    else
+      %{"version" => 1, "files_map" => source_map} -> {:ok, source_map}
+    end
+  end
+
+  @spec get_source_code_map_from_persistent_term() :: source_map() | nil
+  def get_source_code_map_from_persistent_term do
+    :persistent_term.get(@source_code_map_key, nil)
+  end
+
   @spec load_files([Path.t()]) :: source_map()
-  def load_files(paths) when is_list(paths) do
+  def load_files(paths \\ Config.root_source_code_paths()) when is_list(paths) do
     path_pattern = Config.source_code_path_pattern()
     exclude_patterns = Config.source_code_exclude_patterns()
 
@@ -73,7 +129,7 @@ defmodule Sentry.Sources do
     end)
   end
 
-  defp find_files_for_root_path(root_path,path_pattern, exclude_patterns) do
+  defp find_files_for_root_path(root_path, path_pattern, exclude_patterns) do
     root_path
     |> Path.join(path_pattern)
     |> Path.wildcard()